chore: remove unnecessary multistepform variants
